Move slices under slices folder
