+++ src/components/UserDetailHeader.tsx
@@ -1,6 +1,6 @@
 import { useAppDispatch, useAppSelector } from 'store/store';
 import { Button } from 'components';
-import { followUser, unfollowUser } from 'features/singleUser/singleUserSlice';
+import { followUser, unfollowUser } from 'store/slices/singleUserSlice';
 
 interface Props {
   login: string;
