+++ src/store/slices/loadingSlice.ts
