+++ src/store/slices/searchSlice.ts
